bugfix: correcty colliders for LIQUID
